added individual tests for cifi, rafi, and tefi
